% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_abc_kui.R
\name{abc.step1.kui}
\alias{abc.step1.kui}
\title{Change from regular ABC: For every individual, the kui.similarity between P[x, ] and P[M[x], ]
is evaluated. Mutation and evaluation only happens if this similarity value is above
a random uniform number between 1 and 0}
\usage{
abc.step1.kui(P, G, c)
}
\description{
Change from regular ABC: For every individual, the kui.similarity between P[x, ] and P[M[x], ]
is evaluated. Mutation and evaluation only happens if this similarity value is above
a random uniform number between 1 and 0
}

