% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{adjacency_list}
\alias{adjacency_list}
\title{Adjacency List}
\usage{
adjacency_list(G)
}
\arguments{
\item{G}{graph, list with two elements: V - number of nodes in the graph, and E - an edge list in the shape of a 2xE matrix}
}
\value{
The adjacency list of G
}
\description{
Return the adjacency list of a given graph G.
}
