% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_dsatur.R
\name{adjacent_color_set}
\alias{adjacent_color_set}
\title{Auxiliar function for DSatur
Given a set of colors of G, compute adjancent colors to each vertex
This is an auxiliary struture to compute the saturation degrees of G}
\usage{
adjacent_color_set(s, G)
}
\arguments{
\item{s, }{set of colors, size |V|}
}
\value{
a matrix (col: vertices, row: colors) containing the colors adjacent to each vertex
(a V vector of 1:3) and the total number of evaluations spent
}
\description{
Auxiliar function for DSatur
Given a set of colors of G, compute adjancent colors to each vertex
This is an auxiliary struture to compute the saturation degrees of G
}
