% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{connectivity}
\alias{connectivity}
\title{Connectivity}
\usage{
connectivity(V, E)
}
\arguments{
\item{V}{number of nodes}

\item{E}{edge list represented as an |E|x2 matrix}
}
\value{
Boolean whether the graph is connected
}
\description{
Tests a graph for connectivity
}
