% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_dsatur.R
\name{dsatur}
\alias{dsatur}
\title{DSatur (doesn't color all vertices, doesn't generate violation)}
\usage{
dsatur(G, nfe, weight)
}
\description{
DSatur (doesn't color all vertices, doesn't generate violation)
}
