% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Euclidean Distance}
\usage{
euclidean_distance(a, b)
}
\arguments{
\item{a, b}{vectors with same dimension}
}
\value{
Double, the eucliden distance between vector a and vector b
}
\description{
Compute the euclidean distance between two vectors (same dimension).
}
