% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_ffa.R
\name{ffa.improve}
\alias{ffa.improve}
\title{Improve (step2)}
\usage{
ffa.improve(p, w, v, satur, G)
}
\arguments{
\item{p}{vector of colors}

\item{w}{vector of weights}

\item{v}{number of violations}

\item{G}{graph}
}
\value{
list with the permutation of w and its related p, v and number of its functions evaluated
}
\description{
Tries to improve a solution by local search (heuristical swap)
Stops when improvement is detected*
}
\details{
NOTE: The paper is a little bit unclear about this procedure. It says the heuristical swap operator is executed
until the improvements are detected, but it doens't say exactly how to proceed after a no improvement
Some questions of when there is no improvement:
should the next uncolored vertex be considered? or
should the next swap be on the result of the previous swap?

This version assumes the next uncolored vertex should be considered in case of no improvement.*
}
