% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_culberson.R
\name{graph_culberson_variability}
\alias{graph_culberson_variability}
\title{3-Colorable with Variability Graph (+IID)
Source: http://webdocs.cs.ualberta.ca/~joe/Coloring/Colorsrc/index.html}
\usage{
graph_culberson_variability(N, probability = 1, delta = 0)
}
\arguments{
\item{N}{Integer - number of vertices in the graph.}

\item{probability}{0..1 - edge's probability.}

\item{delta}{0, 1, 2 - variability. It generates an uniform graph when delta = 0.}
}
\value{
a list with two elements: V - number of nodes in the graph, and E - an edge list in the shape of a 2xE matrix. The
graph is guaranteed to have a valid 3-coloring, but not guaranteed to be connected.
}
\description{
The algorithm is defined in 3 steps:
}
\details{
1- Create a permutation of nodes
2- Create 3 groups with N/3 nodes, with variability delta
2- Assign edges using IID method
}
\examples{
graph_culberson_variability(10, 1, 0)
}
