% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_minton.R
\name{graph_minton}
\alias{graph_minton}
\title{Generates a 3-GCP instance following the algorithm described by Minton et al.}
\usage{
graph_minton(N, density = 2)
}
\arguments{
\item{N}{Integer - number of vertices in the graph.}

\item{density}{0..1 - edge density in the graph.}
}
\value{
a list of pair arrays representing edges in the graph. The graph is simple and undirected.
}
\description{
S. Minton, M. Johnston, A. Philips, P. Laird, "Minimizing Conflicts: a heuristic repair method
for constraint satisfaction and scheduling problems", Artificial Intelligence, v.58 pp.161-205, 1992
}
\details{
The algorithm is defined in three steps:

1- Create 3 groups with N/3 nodes
2- Create E edges between the groups randomly
3- Accept if the graph has no unconnected components
}
\examples{
generate_minton(10,2)
}

