% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_random.R
\name{graph_random}
\alias{graph_random}
\title{Randomly generates a 3-GCP instance}
\usage{
graph_random(N, density = 2)
}
\arguments{
\item{N}{Integer - number of vertices in the graph.}

\item{density}{0..1 - edge density in the graph.}
}
\value{
a list of pair arrays representing edges in the graph. The graph is simple and undirected.
}
\description{
Randomly generates a 3-GCP instance
}
\examples{
generate_random(10,2)
}

