% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{random_coloring}
\alias{random_coloring}
\title{Random coloring generator}
\usage{
random_coloring(N)
}
\arguments{
\item{N}{Number of vertices}
}
\value{
A vector with N integers 1-3
}
\description{
Generate a random solution for the graph 3-coloring problem.
}

