% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{random_weight}
\alias{random_weight}
\title{Random weight generator}
\usage{
random_weight(N, lb, ub)
}
\arguments{
\item{N}{Number of vertices}

\item{lb}{The upper bound of the values in the weight set (0 <= lv <= ub)}

\item{ub}{The lower bound of the values in the weight set (0 <= lv <= ub)}
}
\value{
A vector with N values (double) in the interval [ub,lb]
}
\description{
Generate a random set of weights for heuristics which use real-values search space.
}
