% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rlevy}
\alias{rlevy}
\title{Random Levy}
\usage{
rlevy(N, b)
}
\arguments{
\item{N}{number of samples to generate}

\item{b}{is a scaling parameter between 0.3 and 1.99}
}
\value{
N samples taken from the Levy Flight Distribution
}
\description{
Levy Flight distribution function, based on
the description by Xin-She Yang,
"Nature-Inspired Metaheuristic Algorithms"
 Second Edition, Luniver Press, (2010).
}

