% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{similarity}
\alias{similarity}
\title{Similarity}
\usage{
similarity(x1, x2)
}
\description{
Calculate the hamming distance-based similarity between two
(sets of) coloring candidates.
}
\details{
Led d be the number of positions with different colors between
coloring x1 and x2, then the similarity between x1 and x2 is
given by:

1 - d / (length of x)
}
