% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_ffa.R
\name{solver_ffa}
\alias{solver_ffa}
\title{Firefly Solver (Fister Version)}
\usage{
solver_ffa(G, nfe, args)
}
\arguments{
\item{G}{the graph to be solved, represented by a list where G$V is the
number of nodes, and G$E is a |E|x2 matrix of edges.}

\item{nfe}{the number of function evaluations. The solver will stop after this
number has been exceeded.}

\item{args}{a list with arguments for the method. The list must contain the
following names:

\itemize{
\item \emph{pop}: Integer > 0. The size of the solution set X
\item \emph{lb}: Number, 0 <= lv <= ub. The lower bound of the values in the weight set
\item \emph{ub}: Number, 0 <= lv <= ub. The upper bound of the values in the weight set
\item \emph{alpha}: Integer > 0. The size of the randomness move withim the search space
\item \emph{beta}: Integer > 0. The attractiveness at distance 0
\item \emph{gamma}: Integer > 0. The absorbtion coefficient
\item \emph{swap}: Boolean. If it is to use the heuristical swap.
}}
}
\value{
A list with three names:
\itemize{
\item{violation}: the number of graph coloring violations of the best solution found (0 for a correct solution)
\item{best}: a vector with the best solution found
\item{evals}: the number of evaluations used by the time the solver stopped.
}
}
\description{
Solves a instance of the 3-GCP problem using the Firefly Algorithm (FFA)
implementation described in Fister et al., 2012.
}
\references{
I. Fister Jr., X.-S. Yang, I. Fister, J. Brest, Memetic firefly algorithm for combinatorial
optimization, in Bioinspired Optimization Methods and their Applications (BIOMA 2012),
B. Filipic and J.Silc, Eds. Jozef Stefan Institute, Ljubljana, Slovenia, 2012
}
