% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{tester.batch}
\alias{tester.batch}
\title{Tester Batch}
\usage{
tester.batch(solverpar.list, graphpar.list, results.file, graphs.file, nfe,
  random.seed = 42, ncores = Inf, output = "none")
}
\arguments{
\item{solverpar.list}{a list where each element is a parameter list for test.solver}

\item{graphpar.list}{a list where each element is a parameter list for dataset}

\item{results.file}{a filename for saving partial and final results}

\item{graphs.file}{a filename for saving partial and final data sets}

\item{nfe}{number of maximum function evaluations for this experiment}

\item{random.seed}{the random number seed for this experiment}

\item{ncores}{the maximum number of cores to be used in this experiment}
}
\value{
Nothing, but saves the files "results.file" and "graphs.file" with the calcultion results
}
\description{
Solves a large number of 3-gcp problems, with a specified set of solver and solver parameters,
and a specified set of graph parameters. Results are saved as the tester goes along, and it is
able to load previous results to avoid re-computation.
}
\examples{
graphpar.list = list()
graphpar.list[["random90"]] = list(name = "random90", size = 5, nodes = 90, density = 3, method = "random")
graphpar.list[["minton90"]] = list(name = "random90", size = 5, nodes = 90, density = 3, method = "minton")

solverpar.list = list()
solverpar.list[["cs.toda"]] = list(name = "cs.toda", method = "cuckoo", pop = 10, pc = 0.0001, compare = F, policy = "levy", alpha = 1, beta = 1.5, E = NULL)
solverpar.list[["abc.togashi"]] = list(name = "abc.togashi", method = "abc", pop = 500, onlooker = 500, c = 3, scout = 1, limit = 300)

tester.batch(solverpar.list, graphpar.list, results.file = "foo.Rda", graphs.file = "bar.Rda", nfe = 10000, random.seed = 42, ncores = 1)

file.remove("foo.test")
file.remove("bar.test")

}
