% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tester.R
\name{tester}
\alias{tester}
\title{Tester}
\usage{
tester(pset, nfe, solverpar, cl = NULL, output = "none")
}
\arguments{
\item{nfe}{the maximum number of evaluations per graph}

\item{solverpar}{a list with the parameters for solver}

\item{problemset}{a list of 3GCP graphs}

\item{solver}{the solver method to be used}
}
\value{
a list with summarized results
}
\description{
Solves a set of 3GCP problems using a specific solving method and returns the results
}
\details{
example of multi-core operation using the parallel library

no_cores <- parallel::detectCores() - 1
cl <- parallel::makeCluster(no_cores)
x <- tester(pset, nfe, solver, NULL, cl = cl)
parallel::stopCluster(cl)
}
