% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_random.R
\name{graph_random}
\alias{graph_random}
\title{Generates a random, valid 3-GCP instance, without connectivity guarantee}
\usage{
graph_random(N, density = 2)
}
\arguments{
\item{N}{Integer - number of vertices in the graph.}

\item{density}{0..1 - edge density in the graph.}
}
\value{
a list with two elements: V - number of nodes in the graph, and E - an edge list in the shape of a 2xE matrix. The
graph is guaranteed to have a valid 3-coloring, but not guaranteed to be connected.
}
\description{
Generates a random, valid 3-GCP instance, without connectivity guarantee
}
\examples{
generate_random(10,2)
}
